// WEKEE.FE/UI.HOME/src/scenes/Admin/MenuSliderAdmin/types.ts
import { ActionType } from 'typesafe-actions';
import * as actions from './actions';

export interface MenuSliderAdminState { // day
    readonly loading: boolean;
    readonly completed: boolean;
    readonly pageIndex: number;
    readonly pageSize: number;
    readonly totalCount: number;
    readonly totalPages: number;
}

export type MenuSliderAdminActions = ActionType<typeof actions>; // day